Prod-44: Break down home controller into supplier and inventory updater controllers
